requirements.txt
openai
pinecone-client
pandas
tqdm
Flask